refactor system as nbl package
