+++ .gitignore
@@ -122,4 +122,3 @@ nbl-gdrive
 data-20_21
 .vscode/settings.json
 test*
-games_*.py
